requirements.txt
streamlit
transformers
torch
watchdog
langchain_huggingface
langchain-community
python-dotenv
openai